requirements.txt
asgiref==3.5.1
crispy-bootstrap5==0.6
Django==3.2.13
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-markitup==4.0.0
Markdown==3.3.7
Pillow==9.1.0
python-dotenv==0.20.0
pytz==2022.1
sorl-thumbnail==12.8.0
sqlparse==0.4.2
